feat(redmine 1246945): add menu style and fix min icon size
