+++ packages/react-front-kit/tsconfig.json
@@ -1,3 +1,3 @@
 {
-  "extends": "tsconfig/react-library.json"
+  "extends": "tsconfig/react-library.json",
 }
