+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
   "extends": "tsconfig/react-library.json",
   "include": [".", ".storybook/*"],
-  "exclude": ["apps", "out", ".husky", ".changeset", ".github"]
+  "exclude": ["apps", "out", ".husky", ".changeset", ".github"],
 }
